requirements.txt
vk_api
telethone
emoji
logging
colorama
json